Create README.md
This is a setup program for COM RS-232C application.
Down load publish folder On desk top .and open publish folder. and install the program with setup.exe 
After install. Click application.  application filename is "WindowsApp3usefulCOMportnov10.application"
When Transciver is in the splitmode .When the keypress down then rescieve split frequency .When the key is released then rescieve main frequency.